refactor(physics): the Python engine must be importable without the web app
The cross-runtime parity gate added in #651 could not actually run. Importing
`backend.alchm_kitchen.main` to reach `compute_kalchm_monica` pulls in fastapi,
sqlalchemy and pyswisseph, so pytest could not even COLLECT the module:

    ModuleNotFoundError: No module named 'fastapi'

A gate that cannot be collected reads as a pass, which is the exact failure mode
this programme keeps finding — a green check that verified nothing.

Extracted the engine to `backend/alchm_kitchen/thermodynamics.py`: pure
arithmetic, `import math` and nothing else. main.py imports from it, so there is
still exactly ONE implementation in the Python runtime. The test now imports the
engine rather than the application.

VERIFIED, from a clean checkout of master in a sibling worktree (jest ignores
/.claude/, which is why this could not be confirmed before merge):

  jest  src/__tests__/kalchmCrossRuntimeParity.test.ts   PASS
        src/__tests__/kalchmFloorDivergence.test.ts      PASS
        47 tests in those two; 162 across all 9 kalchm/monica suites
  pytest backend/tests/test_kalchm_parity.py             41 passed

Both new suites confirmed present in --listTests, not merely "no failures".

CI now runs the pytest half in the static job, installing ONLY pytest. That is
deliberate: if the step ever needs fastapi to collect, the engine has been
re-entangled with the web app and the gate is one refactor from being skipped.
