+++ .github/workflows/monica-integrity.yml
@@ -70,6 +70,29 @@ jobs:
         # neither of the others can see it.
         run: bun scripts/checkNoStrayKalchmFormula.ts
 
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+
+      - name: Cross-runtime engine parity (Python half)
+        # The kalchm formula exists once PER RUNTIME, and the AST gate above is a
+        # TypeScript gate — it structurally cannot see the Python engine, which
+        # serves production traffic via backend/Dockerfile's uvicorn CMD.
+        #
+        # Both halves read the SAME golden vectors
+        # (backend/tests/kalchm_golden_vectors.json), so a change to either
+        # runtime that is not mirrored in the other fails here.
+        #
+        # Only pytest is installed, deliberately: the engine module has NO
+        # framework imports, and keeping it that way is the property being
+        # protected. If this step ever needs fastapi to collect, the engine has
+        # been re-entangled with the web app and the gate is one refactor away
+        # from being skipped.
+        run: |
+          python -m pip install --quiet pytest
+          python -m pytest backend/tests/test_kalchm_parity.py -q
+
   # ---------------------------------------------------------------------------
   # Data. master + cron only: it needs production credentials, so it must never
   # run on a pull_request from a fork.
